+++ src/main/kotlin/org/jmailen/gradle/kotlinter/SourceSetApplier.kt
@@ -1,0 +1,11 @@
+package org.jmailen.gradle.kotlinter
+
+import org.gradle.api.Project
+import org.gradle.api.file.FileTree
+import org.gradle.api.provider.Provider
+
+internal typealias SourceSetAction = (String, Provider<FileTree>) -> Unit
+
+interface SourceSetApplier {
+    fun applyToAll(project: Project, action: SourceSetAction)
+}
